requirement.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
constantly==15.1.0
cryptography==3.1
cymem==2.0.3
daphne==2.5.0
Django==3.1.1
djangorestframework==3.11.1
en-core-web-sm==2.3.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
mathparse==0.1.2
murmurhash==1.0.2
numpy==1.19.2
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==3.13
requests==2.24.0
service-identity==18.1.0
six==1.15.0
spacy==2.3.2
SQLAlchemy==1.3.19
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.1
tqdm==4.49.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
wasabi==0.8.0
whitenoise==5.2.0
zipp==3.1.0
zope.interface==5.1.0
